Core/Random: Changed random functions returning doubles to return floats
* They were all cast to float at use anyway
* Improves roll_chance_f performance (rand32() is now called internally by uniform_real_distribution once instead of twice)

+++ src/server/scripts/ExilesReach/zone_exiles_reach.cpp
@@ -359,7 +359,7 @@ struct npc_sparring_partner_exiles_reach : public ScriptedAI
         {
             case POSITION_SPARPOINT_ADVERTISMENT:
                 me->SetWalk(true);
-                me->GetMotionMaster()->MovePoint(POSITION_SPARPOINT_READY, me->GetFirstCollisionPosition(2.0f, (float)rand_norm() * static_cast<float>(2 * M_PI)));
+                me->GetMotionMaster()->MovePoint(POSITION_SPARPOINT_READY, me->GetFirstCollisionPosition(2.0f, rand_norm() * static_cast<float>(2 * M_PI)));
                 break;
             case POSITION_SPARPOINT_READY:
                 if (Player* player = ObjectAccessor::GetPlayer(*me, _playerGUID))
